refactor: rename file to follow camel casing
